Add tests
